Add files via upload
I've attached the newest version of my C# Snake Game, where I revisited this old project, fixed many bugs and errors, improved the game with new features, and focused effort on the UI/UX experience. I also added a zip file that includes all of the necessary downloads, folders, and files needed for this, so that anyone who downloads this won't have to do anything other than download the zip file and utilize it if they choose. However, I have added just the C# file so that a user does not have to download the zip file if they don't want to.