Add files via upload
This Zip folder replaces the previous one that held the original C# Snake project. This is the newest and updated ZIP folder with the files and the project folders, including library downloads, prerequisites, etc. This folder is designed so that a user can download and run the program with no additional configurations needed.